feat: Add related posts section to post page
This commit adds a related posts section to the post page. It displays a list of other posts that are related to the current post, allowing users to easily navigate to similar content. This feature improves the user experience by providing additional relevant content and encouraging further exploration of the website.